+++ kudosx/cli.py
@@ -9,6 +9,7 @@ from kudosx.commands.explore import explore
 from kudosx.commands.init import init_project
 from kudosx.commands.list import list_skills
 from kudosx.commands.remove import remove
+from kudosx.commands.repo import repo
 from kudosx.commands.search import search
 from kudosx.commands.software import software
 from kudosx.commands.update import update
@@ -30,6 +31,7 @@ cli.add_command(explore)
 cli.add_command(init_project)
 cli.add_command(list_skills)
 cli.add_command(remove)
+cli.add_command(repo)
 cli.add_command(search)
 cli.add_command(software)
 cli.add_command(update)